+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# rversions development version
+
 # rversions 2.1.1
 
 No user visible changes.