+++ NEWS.md
@@ -1,7 +1,7 @@
 
 # rversions development version
 
-* More function to get older versions of R: `r_oldrel_1()`, `r_oldrel_2()`, `r_oldrel_3()`, `r_oldrel_4()`.
+* More function to get older versions of R: `r_oldrel_1()`, `r_oldrel_2()`, `r_oldrel_3()`, `r_oldrel_4()` (@patilindrajeets, #29).
 
 # rversions 2.1.1