+++ NEWS.md
@@ -1,5 +1,7 @@
 
-# rversions development version
+# rversions 2.1.2
+
+No user visible changes.
 
 # rversions 2.1.1